Fixed #112
*. Deleted files for #66
*. Updated Copyright statement mailto:

+++ lib-spiflashstore/src/storemotors.cpp
@@ -2,7 +2,7 @@
  * @file storemotors.cpp
  *
  */
-/* Copyright (C) 2019-2020 by Arjan van Vught mailto:info@raspberrypi-dmx.nl
+/* Copyright (C) 2019-2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -60,7 +60,7 @@ StoreMotors::StoreMotors(void) {
 	s_pThis = this;
 
 	DEBUG_PRINTF("%p", s_pThis);
-	DEBUG_PRINTF("STORE_MOTORS_STRUCT_SIZE=%d", STORE_MOTORS_STRUCT_SIZE);
+	DEBUG_PRINTF("STORE_MOTORS_STRUCT_SIZE=%d", (int) STORE_MOTORS_STRUCT_SIZE);
 
 	assert(STORE_MOTORS_STRUCT_SIZE <= STORE_MOTORS_MAX_SIZE);
