+++ lib-spiflashstore/src/storesparkfundmx.cpp
@@ -2,7 +2,7 @@
  * @file storesparkfundmx.cpp
  *
  */
-/* Copyright (C) 2019 by Arjan van Vught mailto:info@raspberrypi-dmx.nl
+/* Copyright (C) 2019-2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -51,7 +51,7 @@ StoreSparkFunDmx::StoreSparkFunDmx(void) {
 
 	DEBUG_PRINTF("%p", s_pThis);
 	DEBUG_PRINTF("sizeof(TSparkFunDmxParams)=%d", (int) sizeof(struct TSparkFunDmxParams));
-	DEBUG_PRINTF("STORE_SPARKFUN_STRUCT_OFFSET+STORE_SPARKFUN_STRUCT_SIZE=%d", STORE_SPARKFUN_STRUCT_OFFSET + STORE_SPARKFUN_STRUCT_SIZE);
+	DEBUG_PRINTF("STORE_SPARKFUN_STRUCT_OFFSET+STORE_SPARKFUN_STRUCT_SIZE=%d", (int) (STORE_SPARKFUN_STRUCT_OFFSET + STORE_SPARKFUN_STRUCT_SIZE));
 
 	assert(sizeof(struct TSparkFunDmxParams) <= STORE_SPARKFUN_STRUCT_OFFSET);
 	assert((STORE_SPARKFUN_STRUCT_OFFSET + STORE_SPARKFUN_STRUCT_SIZE) <= STORE_SPARKFUN_MAX_SIZE);
